requirements.txt
linkedin-jobs-scraper==5.0.2
google-genai>=1.27.0
google-generativeai>=0.8.5